Fix Redis & Tair vector store overwrite newly created vector
We should not overwrite the vectors for nodes batches that
come after the first iteration.

+++ llama-index-integrations/vector_stores/llama-index-vector-stores-redis/llama_index/vector_stores/redis/base.py
@@ -165,7 +165,7 @@ class RedisVectorStore(BasePydanticVectorStore):
         self._index_args["dims"] = len(nodes[0].get_embedding())
 
         if self._index_exists():
-            if self._overwrite:
+            if add_kwargs.get("first_batch", True) and self._overwrite:
                 self.delete_index()
                 self._create_index()
             else:
