+++ llama-index-integrations/vector_stores/llama-index-vector-stores-tair/llama_index/vector_stores/tair/base.py
@@ -139,7 +139,7 @@ class TairVectorStore(VectorStore):
         self.dim = len(nodes[0].get_embedding())
 
         if self._index_exists():
-            if self._overwrite:
+            if add_kwargs.get("first_batch", True) and self._overwrite:
                 self.delete_index()
                 self._create_index()
             else:
